requirements.txt
# Core AI
anthropic>=0.40.0

# MCP (Model Context Protocol)
mcp[cli]>=1.0.0

# Vector Database
pinecone>=3.0.0

# MySQL Driver
mysql-connector-python>=8.0.0

# Environment Variables
python-dotenv>=1.0.0

# Existing dependencies
numpy>=1.24.0
flask>=3.0.0
